Added overwrite kwarg to download_file

+++ model_toolkit/utils/io.py
@@ -14,7 +14,7 @@ def load_pickle(path: Path):
         return pickle.load(f)
 
 
-def download_file(url, output_dir='', output_filename=None):
+def download_file(url, output_dir='', output_filename=None, overwrite=False):
     """Downloads a file to the filesystem.
     
     Args:
@@ -23,6 +23,8 @@ def download_file(url, output_dir='', output_filename=None):
         output_filename: a `str`, what to name the downloaded file.
             If this value is 'None', the filename is extracted from the URL.
             Default is 'None'.
+        overwrite: a `bool`, indicates whether to overwrite the file if 
+            it already exists, default is 'False'
     """
     output_dir = Path(output_dir)
     os.makedirs(output_dir, exist_ok=True)
@@ -30,6 +32,10 @@ def download_file(url, output_dir='', output_filename=None):
     output_filename = output_filename or url.split('/')[-1]
     output_filepath = output_dir / output_filename
     
+    if output_filepath.exists() and not overwrite:
+        print(f"File {str(output_filepath)} already exists, nothing to do.")
+        return
     bytestream = urllib.request.urlopen(url)
     with open(output_filepath, 'w+b') as output_file:
         output_file.write(bytestream.read())
+        print(f"File saved to {str(output_filepath)}.")
